* debian/control (Build-Depends): Depend on debhelper 8.9.7 for using   "override-*-arch" and "override-*-indep" targets in debian/rules. * debian/rules: Completely rewritten to use features from debhelper v7. * plugins.d/mandos-client.c (fprintf_plus): Cast return value to   return type; this fixes a warning since TEMP_FAILURE_RETRY seems to   return a long int, not whatever type its inner expresson returns. * plugins.d/password-prompt.c (fprintf_plus): - '' -

+++ plugins.d/mandos-client.c
@@ -187,7 +187,7 @@ int fprintf_plus(FILE *stream, const char *format, ...){
   
   TEMP_FAILURE_RETRY(fprintf(stream, "Mandos plugin %s: ",
 			     program_invocation_short_name));
-  return TEMP_FAILURE_RETRY(vfprintf(stream, format, ap));
+  return (int)TEMP_FAILURE_RETRY(vfprintf(stream, format, ap));
 }
 
 /*
